Introduce pseudo-FLASH storage on top of stack
This storage has a limitation that it isn't possible to write a the single word twice
